Zero downtime upgrades, merge-based index construction

+++ code/features-index/index-query/src/main/java/nu/marginalia/index/query/IndexQueryBuilder.java
@@ -9,16 +9,16 @@ import nu.marginalia.index.query.filter.QueryFilterStepIf;
 public interface IndexQueryBuilder {
     /** Filters documents that also contain termId, within the full index.
      */
-    IndexQueryBuilder alsoFull(int termId);
+    IndexQueryBuilder alsoFull(long termId);
 
     /**
      * Filters documents that also contain the termId, within the priority index.
      */
-    IndexQueryBuilder alsoPrio(int termIds);
+    IndexQueryBuilder alsoPrio(long termIds);
 
     /** Excludes documents that contain termId, within the full index
      */
-    IndexQueryBuilder notFull(int termId);
+    IndexQueryBuilder notFull(long termId);
 
     IndexQueryBuilder addInclusionFilter(QueryFilterStepIf filterStep);
